src/pyproject.toml
[tool.poetry]
name = "hanusovedni"
version = "0.0.1"
description = ""
authors = ["jozo <hi@jozo.io>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
wagtail = "^2.8"
uwsgi = "^2.0.18"
psycopg2 = "^2.8.4"
wagtail-autocomplete = "^0.5"
sentry-sdk = "^0.14.1"
django-redis = "^4.11.0"
django-debug-toolbar = "^2.2"
requests = "^2.23.0"
cairosvg = "^2.4.2"

[tool.poetry.dev-dependencies]
django-extensions = "^2.2.6"
black = "^19.10b0"
ipython = "^7.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
